requirements.txt
ccxt==4.1.24
websockets==11.0.3
requests==2.31.0

pandas==2.2.1
numpy==1.26.4

pymongo==4.6.1
redis==5.0.1
flask


asyncio==3.4.3
schedule==1.2.0
fastapi==0.110.0
uvicorn==0.27.1
prometheus-client==0.20.0

python-dotenv==1.0.1

pytest==8.0.2
pytest-asyncio==0.23.6
pytest-mock
mongomock==4.1.2
requests-mock==1.11.0

loguru==0.7.2

psutil==5.9.8
